+++ shared/src/commonMain/kotlin/com/vickikbt/devtyme/domain/repositories/SummariesRepository.kt
@@ -7,7 +7,7 @@ interface SummariesRepository {
 
     suspend fun fetchSummaries(start: String? = null, range: String? = null): Flow<Summaries?>
 
-    suspend fun saveDailyGoal(hours: Int)
+    suspend fun saveDailyGoal(hours: Long)
 
-    suspend fun getDailyGoal(): Flow<Int?>
+    suspend fun getDailyGoal(): Flow<Long?>
 }
